+++ components/shared/header/menu.tsx
@@ -7,14 +7,20 @@ const Menu = () => {
     <div className="flex justify-end gap-3">
       <nav className="hidden md:flex w-full max-w-x gap-1">
         <Button asChild variant="ghost">
-          <Link href="/visual">Visualization</Link>
+          <Link href="/home">Home</Link>
         </Button>
         <Button asChild variant="ghost">
+          <Link href="/portfolio">Portfolio</Link>
+        </Button>
+        <Button asChild variant="ghost">
+          <Link href="/community">Community</Link>
+        </Button>
+        {/* <Button asChild variant="ghost">
           <Link href="/about">About Us</Link>
         </Button>
         <Button asChild variant="ghost">
           <Link href="/cart">Our Impact</Link>
-        </Button>
+        </Button> */}
         <Button asChild>
           <Link href="/auth">Login</Link>
         </Button>
